fix: remove view mode dependency
