+++ packages/frontend/src/components/Explorer/ExplorerHeader/index.tsx
@@ -1,69 +1,13 @@
-import { Button, Classes } from '@blueprintjs/core';
-import { Tooltip2 } from '@blueprintjs/popover2';
-import React, { FC, useState } from 'react';
-import { useUpdateMutation } from '../../../hooks/useSavedQuery';
-import { useExplorer } from '../../../providers/ExplorerProvider';
-import { TrackSection } from '../../../providers/TrackingProvider';
-import { SectionName } from '../../../types/Events';
-import { UpdatedInfo } from '../../common/ActionCard';
+import React, { FC } from 'react';
 import { RefreshButton } from '../../RefreshButton';
-import RefreshServerButton from '../../RefreshServer';
-import RenameSavedChartModal from '../../SavedQueries/RenameSavedChartModal';
-import { ChartName, TitleWrapper, Wrapper } from './ExplorerHeader.styles';
+import RefreshDbtButton from '../../RefreshDbtButton';
+import { Wrapper } from './ExplorerHeader.styles';
 
-const ExplorerHeader: FC = () => {
-    const mode = 'view';
-    const {
-        state: { savedChart },
-    } = useExplorer();
-    const [isRenamingChart, setIsRenamingChart] = useState(false);
-    const updateSavedChart = useUpdateMutation(savedChart?.uuid);
-    const isViewMode = false;
-
-    return (
-        <TrackSection name={SectionName.EXPLORER_TOP_BUTTONS}>
-            <Wrapper>
-                <TitleWrapper>
-                    {savedChart && (
-                        <>
-                            <ChartName
-                                className={Classes.TEXT_OVERFLOW_ELLIPSIS}
-                            >
-                                {savedChart.name}
-                                <UpdatedInfo
-                                    updatedAt={savedChart.updatedAt}
-                                    user={savedChart.updatedByUser}
-                                />
-                            </ChartName>
-                            {savedChart.description && (
-                                <Tooltip2
-                                    content={savedChart.description}
-                                    position="bottom"
-                                >
-                                    <Button icon="info-sign" minimal />
-                                </Tooltip2>
-                            )}
-                            {!isViewMode && (
-                                <Button
-                                    icon="edit"
-                                    disabled={updateSavedChart.isLoading}
-                                    onClick={() => setIsRenamingChart(true)}
-                                    minimal
-                                />
-                            )}
-                            <RenameSavedChartModal
-                                savedChartUuid={savedChart.uuid}
-                                isOpen={isRenamingChart}
-                                onClose={() => setIsRenamingChart(false)}
-                            />
-                        </>
-                    )}
-                </TitleWrapper>
-                <RefreshButton />
-                <RefreshServerButton />
-            </Wrapper>
-        </TrackSection>
-    );
-};
+const ExplorerHeader: FC = () => (
+    <Wrapper>
+        <RefreshDbtButton />
+        <RefreshButton />
+    </Wrapper>
+);
 
 export default ExplorerHeader;
