--- packages/frontend/src/components/Explorer/RunButtonsSection/RunButtonsSection.styles.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled.div`
-    height: 60px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    button {
-        margin: 0;
-    }
-`;
