+++ packages/frontend/src/components/Explorer/index.tsx
@@ -1,16 +1,15 @@
 import { FC } from 'react';
 import { ExplorerWrapper } from './Explorer.styles';
+import ExplorerHeader from './ExplorerHeader';
 import FiltersCard from './FiltersCard/FiltersCard';
 import ResultsCard from './ResultsCard/ResultsCard';
-import RunButtonsSection from './RunButtonsSection';
 import SqlCard from './SqlCard/SqlCard';
 import VisualizationCard from './VisualizationCard/VisualizationCard';
 
-const Explorer: FC<{ isExplore?: boolean }> = ({ isExplore = false }) => {
+const Explorer: FC = () => {
     return (
         <ExplorerWrapper>
-            {/* Remove conditional once the buttons are updated in the header*/}
-            {isExplore && <RunButtonsSection />}
+            <ExplorerHeader />
             <FiltersCard />
             <VisualizationCard />
             <ResultsCard />
